Displays Entered Number as Sum of Two Prime No.
This program takes a positive integer value as an input from the user and then displays the integer as the sum of two prime numbers. If the number cannot be expressed as the sum of two prime numbers then displays the message of "Sorry. Can't be expressed as the sum of primes."